Move metrics infra into node-common. Move mempool metrics under mempool/
