+++ core-rust/state-manager/src/mempool/metrics.rs
@@ -62,9 +62,77 @@
  * permissions under this License.
  */
 
-pub mod config;
-pub mod environment;
-pub mod java;
-pub mod jni;
-pub mod metrics;
-pub mod utils;
+use node_common::metrics::*;
+use prometheus::*;
+
+use crate::{MempoolAddError, MempoolAddSource, RejectionReason};
+
+pub struct MempoolMetrics {
+    pub current_transactions: IntGauge,
+    pub current_total_transactions_size: IntGauge,
+    pub submission_added: IntCounterVec,
+    pub submission_rejected: IntCounterVec,
+    pub from_local_api_to_commit_wait: Histogram,
+}
+
+impl MempoolMetrics {
+    pub fn new(registry: &Registry) -> Self {
+        Self {
+            current_transactions: IntGauge::with_opts(opts(
+                "mempool_current_transactions",
+                "Number of transactions in progress in the mempool.",
+            )).registered_at(registry),
+            current_total_transactions_size: IntGauge::with_opts(opts(
+                "mempool_current_total_transactions_size",
+                "Total size in bytes of transactions in mempool.",
+            )).registered_at(registry),
+            submission_added: IntCounterVec::new(
+                opts(
+                    "mempool_submission_added_total",
+                    "Count of submissions added to the mempool.",
+                ),
+                &["source"],
+            ).registered_at(registry),
+            submission_rejected: IntCounterVec::new(
+                opts(
+                    "mempool_submission_rejected_total",
+                    "Count of the submissions rejected by the mempool.",
+                ),
+                &["source", "rejection_reason"],
+            ).registered_at(registry),
+            from_local_api_to_commit_wait: new_timer(
+                opts(
+                    "mempool_from_local_api_to_commit_wait",
+                    "Time spent in the mempool, by a transaction coming from local API, until successful commit."
+                ),
+                vec![0.01, 0.1, 0.5, 1.0, 2.5, 5.0, 10.0, 25.0, 50.0]
+            ).registered_at(registry)
+        }
+    }
+}
+
+impl MetricLabel for MempoolAddSource {
+    type StringReturnType = &'static str;
+
+    fn prometheus_label_name(&self) -> Self::StringReturnType {
+        match *self {
+            MempoolAddSource::CoreApi => "CoreApi",
+            MempoolAddSource::MempoolSync => "MempoolSync",
+        }
+    }
+}
+
+impl MetricLabel for MempoolAddError {
+    type StringReturnType = &'static str;
+
+    fn prometheus_label_name(&self) -> Self::StringReturnType {
+        match self {
+            MempoolAddError::PriorityThresholdNotMet { .. } => "PriorityThresholdNotMet",
+            MempoolAddError::Rejected(rejection) => match &rejection.reason {
+                RejectionReason::FromExecution(_) => "ExecutionError",
+                RejectionReason::ValidationError(_) => "ValidationError",
+            },
+            MempoolAddError::Duplicate(_) => "Duplicate",
+        }
+    }
+}
